[CMake] Revamp CMake for readability
CMake has been a bit of a trash-heap for the many projects I have copied
it to and from. After watching some talks on YouTube it's time I applied
what was said there to these files. This will make the project a lot
easier to understand compared to where it was previously.