Catch exceptions in the log class so we can use it elsewhere.
Catch exceptions in the log class which will allow us to use it in other
places where exceptions will be thrown. We don't want to enter an
infinite loop where the exception is caused by an action from catching
the previous exception. This class can now be treated like a C function
ad simply print info to the terminal.