Write a new HTTP parse class, update log lines.
Wrote a new class to parse the HTTP protocol. This lets us just
interpret the data coming in as a whole and not have to deal with much.
Hopefully this method works (I have yet to run any of this code in the
	whole codebase tbh). Updated some log lines in other parts of
the codebase and really standardized on the log class. Removed most of
the mutex code from the log class as it will cause recursive hangs if
the log class is called in the same thread twice (happened in another
	project of mine). We really only need to lock when we have a
call to write to stdout so we don't write two sentences on top of each
other in the terminal/file (making things unreadable) so only lock a
mutex there. Everything else should be thread safe (but don't quote me).