Made a bunch of stubs for kstring
Added a bunch of stubs for kstring which are needed to make it act like
std::string. These stubs just need to be filled in and it will make
kstring identical to std::string minus the unicode handling. Internally
kstring will handle memory a bit differently, doing reference counting
on the object which should prevent some undefined behavior when copying
strings, this also manages memory better.