Finish out the kstring rewrite -- still does not compile
Finished rewriting the internal structure of kstring, fixed some bugs
and such along the way as well. The constructors and operator= functions
still need to be swapped since C++ is too dumb to understand a reference
to the constructor instead of the object itself. Once this is finished,
I can begin work on creating new functions and doing testing to ensure
everything works the way it should.